Drop normal scheduler
The decision of having a support for more then one scheduler is an implementation detail of this package and not something that should have been exposed to users of this package.
